Refactor: 로그인 시 로딩창 핸들링
