+++ src/Components/Common/CelebratingModal.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import theme from "../../../Styles/Theme";
+import theme from "../../Styles/Theme";
 
 function CelebratingModal({ celebratingMessage }) {
   return (
@@ -32,7 +32,7 @@ const Container = styled.div`
   transform: translate(-50%, -50%);
   box-shadow: -14px -14px 20px rgba(0, 0, 0, 0.02),
     14px 14px 20px rgba(0, 0, 0, 0.05);
-  z-index: 2;
+  z-index: 3;
 `;
 
 const CelebratingImg = styled.div`
